requirements.txt
pandas
jupyter
python-dotenv   # env variables for notebooks
SQLAlchemy
PyMySQL